+++ src/main/components/ArrangeView/ArrangeView.tsx
@@ -20,6 +20,7 @@ import {
 } from "../../actions"
 import { Layout } from "../../Constants"
 import { getClientPos } from "../../helpers/mouseEvent"
+import { observeDrag } from "../../helpers/observeDrag"
 import { useContextMenu } from "../../hooks/useContextMenu"
 import { useStores } from "../../hooks/useStores"
 import { useTheme } from "../../hooks/useTheme"
@@ -27,7 +28,6 @@ import { GLCanvas } from "../GLCanvas/GLCanvas"
 import { HorizontalScaleScrollBar } from "../inputs/ScaleScrollBar"
 import { BAR_WIDTH, VerticalScrollBar } from "../inputs/ScrollBar"
 import CanvasPianoRuler from "../PianoRoll/CanvasPianoRuler"
-import { observeDrag } from "../PianoRoll/MouseHandler/observeDrag"
 import { ArrangeContextMenu } from "./ArrangeContextMenu"
 import { ArrangeTrackContextMenu } from "./ArrangeTrackContextMenu"
 import { ArrangeViewRenderer } from "./ArrangeViewRenderer"
