+++ src/main/components/ControlPane/VelocityControl/VelocityControl.tsx
@@ -5,10 +5,10 @@ import styled from "styled-components"
 import { containsPoint, IPoint, IRect } from "../../../../common/geometry"
 import { isNoteEvent } from "../../../../common/track"
 import { changeNotesVelocity } from "../../../actions"
+import { observeDrag } from "../../../helpers/observeDrag"
 import { useStores } from "../../../hooks/useStores"
 import { useTheme } from "../../../hooks/useTheme"
 import { GLCanvas } from "../../GLCanvas/GLCanvas"
-import { observeDrag } from "../../PianoRoll/MouseHandler/observeDrag"
 import { GraphAxis } from "../Graph/GraphAxis"
 import { VelocityControlRenderer } from "./VelocityControlRenderer"
