+++ src/main/components/PianoRoll/PianoNotes.tsx
@@ -5,10 +5,10 @@ import { useTheme } from "styled-components"
 import { containsPoint, zeroRect } from "../../../common/geometry"
 import { getSelectionBounds } from "../../../common/selection/Selection"
 import { removeEvent } from "../../actions"
+import { observeDoubleClick } from "../../helpers/observeDoubleClick"
 import { useContextMenu } from "../../hooks/useContextMenu"
 import { useStores } from "../../hooks/useStores"
 import { GLCanvas } from "../GLCanvas/GLCanvas"
-import { observeDoubleClick } from "./MouseHandler/observeDoubleClick"
 import PencilMouseHandler from "./MouseHandler/PencilMouseHandler"
 import SelectionMouseHandler from "./MouseHandler/SelectionMouseHandler"
 import { PianoRollRenderer } from "./PianoRollRenderer/PianoRollRenderer"
