+++ src/main/helpers/observeDoubleClick.ts
